+++ house-dashboard/src/lib.rs
@@ -30,12 +30,6 @@ use futures::{future::ready, stream::FuturesUnordered, StreamExt};
 
 use image::{ImageFormat, RgbImage};
 
-#[cfg(target_arch = "x86_64")]
-use extrasafe::{
-    builtins::{danger_zone::ForkAndExec, BasicCapabilities, Networking, SystemIO},
-    SafetyContext,
-};
-
 use isahc::{
     auth::{Authentication, Credentials},
     config::{CaCertificate, Configurable, SslOption},
@@ -64,9 +58,6 @@ use self::logging::setup as setup_logging;
 pub async fn main() -> Result<(), Report> {
     let arguments = parse_command_line();
     setup_logging(arguments.verbosity.try_into().into_diagnostic()?)?;
-
-    #[cfg(target_arch = "x86_64")]
-    setup_allowed_syscalls()?;
 
     let (style_configuration, influxdb_configuration) =
         parse_configuration(&arguments.configuration_directory_path)
@@ -112,48 +103,6 @@ pub async fn main() -> Result<(), Report> {
         .await
         .wrap_err("cannot save image")?;
     }
-
-    Ok(())
-}
-
-/// Setup allowed syscalls
-#[cfg(target_arch = "x86_64")]
-fn setup_allowed_syscalls() -> Result<(), Report> {
-    let safety_context = SafetyContext::new();
-
-    let safety_context = safety_context
-        .enable(BasicCapabilities)
-        .into_diagnostic()
-        .wrap_err("cannot enable basic syscalls")?;
-
-    let safety_context = safety_context
-        .enable(
-            SystemIO::nothing()
-                .allow_read()
-                .allow_write()
-                .allow_ioctl()
-                .allow_metadata()
-                .allow_open()
-                .yes_really()
-                .allow_close(),
-        )
-        .into_diagnostic()
-        .wrap_err("cannot enable system IO syscalls")?;
-
-    let safety_context = safety_context
-        .enable(Networking::nothing().allow_start_tcp_clients())
-        .into_diagnostic()
-        .wrap_err("cannot enable networking syscalls")?;
-
-    let safety_context = safety_context
-        .enable(ForkAndExec)
-        .into_diagnostic()
-        .wrap_err("cannot enable fork and exec syscalls")?;
-
-    safety_context
-        .apply_to_all_threads()
-        .into_diagnostic()
-        .wrap_err("cannot apply safety context")?;
 
     Ok(())
 }