+++ Changelog.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Keep lexicographical order of chart configuration files
 
+### Removed
+
+- Remove syscalls filtering
+
 
 ## [1.1.0] - 2023-12-18